Seguranca/17 implementar laravel policy

+++ app/Livewire/Views/Empresas/Listagem.php
@@ -3,8 +3,11 @@
 namespace App\Livewire\Views\Empresas;
 
 use App\Models\Empresa;
+use App\Models\User;
+use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Pagination\LengthAwarePaginator;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Attributes\Layout;
 use Livewire\Attributes\Title;
 use Livewire\Component;
@@ -27,6 +30,16 @@ class Listagem extends Component
 
     public bool $modalConfirmandoRemocaoEmpresa = false;
 
+    public User|Authenticatable $usuario;
+
+    public function mount(): void
+    {
+        $this->usuario = Auth::user();
+        if ($this->usuario->cannot('viewAny', \App\Models\Empresa::class)) {
+            abort('401', 'Você não tem permissão para acessar essa página');
+        }
+    }
+
     #[Layout('components.layouts.main')]
     #[Title('SAFI NFE - Listagem de Empresas')]
     public function render()
@@ -36,11 +49,19 @@ class Listagem extends Component
 
     public function irCadastrar(): void
     {
+        if ($this->usuario->cannot('create', \App\Models\Empresa::class)) {
+            $this->error('Você não tem permissão para fazer isso.');
+            return;
+        }
         redirect('/empresas/cadastro');
     }
 
     public function excluirEmpresa(): void
     {
+        if ($this->usuario->cannot('delete', $this->empresaAtual)) {
+            $this->error('Você não tem permissão para fazer isso.');
+            return;
+        }
         $this->empresaAtual->endereco()->first()->delete();
         $this->empresaAtual->delete();
 
@@ -51,11 +72,20 @@ class Listagem extends Component
     public function setRemocaoEmpresa(?int $empresa_id): void
     {
         $this->empresaAtual = Empresa::find($empresa_id);
+        if ($this->usuario->cannot('delete', $this->empresaAtual)) {
+            $this->error('Você não tem permissão para fazer isso.');
+            return;
+        }
         $this->modalConfirmandoRemocaoEmpresa = true;
     }
 
-    public function irEdicaoEmpresa(int $empresa_id): Redirector|RedirectResponse
+    public function irEdicaoEmpresa(int $empresa_id): void
     {
-        return redirect("/empresas/edicao/{$empresa_id}");
+        $this->empresaAtual = Empresa::find($empresa_id);
+        if ($this->usuario->cannot('update', $this->empresaAtual)) {
+            $this->error('Você não tem permissão para fazer isso.');
+            return;
+        }
+        redirect("/empresas/edicao/{$empresa_id}");
     }
 }
